fix moderation to integer

+++ general/models.py
@@ -469,7 +469,7 @@ class Assessment(UUIDModel):
     ]
     type = models.CharField(choices=type_choices, max_length=1, default='C')
 
-    moderation = models.DecimalField(max_digits=3, decimal_places=2, default=0.0)
+    moderation = models.IntegerField(default=0)
     moderated_by = models.ForeignKey(User, on_delete=models.SET_NULL, null=True, blank=True)
     moderation_datetime = models.DateTimeField(null=True, blank=True)
